requirements.txt
Django==1.11.16
django-cart
gunicorn==19.0.0
pytz==2018.7
six==1.11.0
dj-database-url==0.5.0 
django-heroku==0.2.1
whitenoise==4.1.1
waitress
dj_static
requests
django-rest-auth==0.9.3
djangorestframework==3.9.0